Broke out logic for popular items to own component
